+++ packages/localize-tools/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## Unreleased -->
 
+## [0.2.1] - 2021-04-02
+
+### Changed
+
+- XLIFF file headers have been simplified to:
+
+```xml
+<xliff version="1.2" xmlns="urn:oasis:names:tc:xliff:document:1.2">
+```
+
 ## [0.2.0] - 2021-03-30
 
 ### Changed